// src/types.d.ts
export type DaDataResult<T> = {
  value: string,
  unrestricted_value: string,
  data: T
}

export type DaDataAddressMetro = {
  name: string;
  line: string;
  distance: number;
}

export type DaDataAddress = {
  area: string | null;
  area_fias_id: string | null;
  area_kladr_id: string | null;
  area_type: string | null;
  area_type_full: string | null;
  area_with_type: string | null;
  beltway_distance: null;
  beltway_hit: null;
  block: string | null;
  block_type: string | null;
  block_type_full: string | null;
  federal_district: string | null;
  capital_marker: '0' | '1' | '2' | '3' | '4';
  city: string | null;
  city_area: string | null;
  city_district: string | null;
  city_district_fias_id: string | null;
  city_district_kladr_id: string | null;
  city_district_type: string | null;
  city_district_type_full: string | null;
  city_district_with_type: string | null;
  city_fias_id: string | null;
  city_kladr_id: string | null;
  city_type: string | null;
  city_type_full: string | null;
  city_with_type: string | null;
  country: string;
  country_iso_code: string;
  fias_id: string;
  fias_level: string;
  flat: string | null;
  flat_area: null;
  flat_price: null;
  flat_type: string | null;
  flat_type_full: string | null;
  geo_lat: string | null;
  geo_lon: string | null;
  geoname_id: string | null;
  history_values: string[] | null;
  house: string | null;
  house_fias_id: string | null;
  house_kladr_id: string | null;
  house_type: string | null;
  house_type_full: string | null;
  kladr_id: string;
  okato: string | null;
  oktmo: string | null;
  postal_box: string | null;
  postal_code: string | null;
  qc: null;
  qc_complete: null;
  qc_geo: '0' | '1' | '2' | '3' | '4' | '5';
  qc_house: null;
  region: string;
  region_fias_id: string;
  region_kladr_id: string;
  region_type: string;
  region_type_full: string;
  region_with_type: string;
  settlement: string | null;
  settlement_fias_id: string | null;
  settlement_kladr_id: string | null;
  settlement_type: string | null;
  settlement_type_full: string | null;
  settlement_with_type: string | null;
  source: string | null;
  square_meter_price: null;
  street: string | null;
  street_fias_id: string | null;
  street_kladr_id: string | null;
  street_type: string | null;
  street_type_full: string | null;
  street_with_type: string | null;
  tax_office: string | null;
  tax_office_legal: string | null;
  timezone: string | null;
  unparsed_parts: null;
  fias_code: string;
  region_iso_code: string;
  fias_actuality_state: string;
  metro: DaDataAddressMetro[] | null;
}